Moved TzevaAdomListener to its own package
Makes it easier to find this core interface + more related classes will be added to the package soon.
