+++ src/main/java/dte/tzevaadomapi/notifier/TzevaAdomNotifier.java
@@ -11,6 +11,7 @@ import java.util.function.Consumer;
 import dte.tzevaadomapi.alert.Alert;
 import dte.tzevaadomapi.alertsource.AlertSource;
 import dte.tzevaadomapi.alertsource.PHOAlertSource;
+import dte.tzevaadomapi.listener.TzevaAdomListener;
 import dte.tzevaadomapi.utils.UncheckedExceptions.CheckedSupplier;
 
 /**
